+++ packages/server/src/lib/config.test.ts
@@ -36,6 +36,7 @@ describe('loadConfig', () => {
       moduleConfigPath: undefined,
       siteName: undefined,
       webBundleDirectory: undefined,
+      regions: [],
     })
   })